chore(backend): javac警告根治(D-102/D-103)——pom固化-Xlint:all防警告默默积累+清零125条存量警告(弃用API替换14/static类名限定30/lossy显式截断10/raw参数化13/varargs与try压制4/死代码5)+6文件IDE警告清理(未用@Autowired字段/StyleStageCompletionHelper死链/text-block尾随空格)；三重验证归零(javac警告0+编译EXIT=0+IDE诊断0)

+++ backend/src/main/java/com/fashion/supplychain/intelligence/agent/planning/AgentPlanningEngine.java
@@ -159,7 +159,7 @@ public class AgentPlanningEngine {
         if (userMessage.length() > 200) score += 10;
 
         long questionCount = lower.chars().filter(c -> c == '?' || c == '？').count();
-        score += Math.min(questionCount * 5, 15);
+        score += (int) Math.min(questionCount * 5, 15);
 
         // 复杂分析类动词
         java.util.regex.Pattern complexPattern = java.util.regex.Pattern.compile(
